+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added elements
   - [CalendarControl](src/controls/calendar.ts) [**demo**](https://yegorich555.github.io/web-ui-pack/control/calendar)
 - Added [helpers](README.md#helpers)
+  - [`dateCopyTime`](src/helpers/dateCopyTime.ts)
   - [`dateFromString`](src/helpers/dateFromString.ts)
   - [`dateToString`](src/helpers/dateToString.ts)
   - [`onScrollStop`](src/helpers/onScrollStop.ts)